# src/main/webapp/swagger-ui/swagger-rmes.yaml
openapi: 3.0.0
servers:
  - url: 'http://localhost:8080/metadata-api'
    description: 'Local dev server'
  - url: 'http://qfrmeswnczlht01'
    description: 'Qf server'
  - url: 'http://dvrmeswnczlht01'
    description: 'Local dev server'
info:
  description: API sur les métadonnées de l'Insee
  version: '1.1.1'
  title: API RMéS
...
paths:
  '/nomenclatures':
    get:
      tags:
        - nomenclatures
      summary: Liste des nomenclatures disponibles (autres que "catégories juridiques")
      operationId: getClassifications
      responses:
        '200':
          description: Opération réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nomenclatures'            
            application/xml:
              schema:
                $ref: '#/components/schemas/Nomenclatures'            
        '400':
          description: La syntaxe de la requête est incorrecte
        '401':
          description: Une authentification est nécessaire pour accéder à la ressource
        '404':
          description: Ressource non trouvée
        '406':
          description: L'en-tête HTTP 'Accept' contient une valeur non acceptée
        '500':
          description: Erreur interne du serveur
  '/nomenclatures/{code}':
    get:
      tags:
        - nomenclatures
      summary: Postes d'une nomenclature (autres que "catégories juridiques")
      operationId: getClassificationByCode
      parameters:
        - name: code
          in: path
          description: Identifiant de la nomenclature (hors cj)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Opération réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Postes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Postes'
        '400':
          description: La syntaxe de la requête est incorrecte
        '401':
          description: Une authentification est nécessaire pour accéder à la ressource
        '404':
          description: Ressource non trouvée
        '406':
          description: L'en-tête HTTP 'Accept' contient une valeur non acceptée
        '500':
          description: Erreur interne du serveur
  '/correspondances':
    get:
      tags:
        - nomenclatures
      summary: Correspondances entre deux nomenclatures
      operationId: getAllCorrespondences
      responses:
        '200':
          description: Opération réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescriptionsTablesCorrespondance'
            application/xml:
              schema:
                $ref: '#/components/schemas/DescriptionsTablesCorrespondance'
        '400':
          description: La syntaxe de la requête est incorrecte
        '401':
          description: Une authentification est nécessaire pour accéder à la ressource
        '404':
          description: Ressource non trouvée
        '406':
          description: L'en-tête HTTP 'Accept' contient une valeur non acceptée
        '500':
          description: Erreur interne du serveur
  '/correspondances/{codeNomenclatureSource}/{codeNomenclatureCible}':
    get:
      tags:
        - nomenclatures
      summary: Correspondances entre deux nomenclatures
      operationId: getCorrespondencesByCodes
      parameters:
        - name: codeNomenclatureSource
          in: path
          description: Identifiant de la nomenclature (source)
          required: true
          schema:
            type: string
        - name: codeNomenclatureCible
          in: path
          description: Identifiant de la nomenclature (cible)
          required: true
          schema:
            type: string            
      responses:
        '200':
          description: Opération réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Correspondances'
            application/xml:
              schema:
                $ref: '#/components/schemas/Correspondances'
        '400':
          description: La syntaxe de la requête est incorrecte
        '401':
          description: Une authentification est nécessaire pour accéder à la ressource
        '404':
          description: Ressource non trouvée
        '406':
          description: L'en-tête HTTP 'Accept' contient une valeur non acceptée
        '500':
          description: Erreur interne du serveur
...
components:
  schemas:
    Poste:
      type: object
      description: Objet représentant un poste d'une nomenclature
      properties:
        code:
          type: string
          example: '23.99Z'
          xml:
            attribute: true
        uri:
          type: string
          example: 'http://id.insee.fr/codes/nafr2/sousClasse/23.99Z'
          xml:
            attribute: true          
        codeParent:
          type: string
          example: '23.99'
          xml:
            attribute: true
        uriParent:
          type: string
          example: 'http://id.insee.fr/codes/nafr2/classe/23.99'
          xml:
            attribute: true
        intituleFr:
          type: string
          example: Fabrication d'autres produits minéraux non métalliques n.c.a.
          xml:
            name : IntituleFr
        intituleEn: 
          type: string
          example: Manufacture of other non-metallic mineral products n.e.c.
          xml:
            name : IntituleEn
        noteGenerale:
          type: string
          example: exemple de contenu général
          xml:
            name : NoteGenerale
        contenuLimite: 
          type: string
          example: exemple contenu limite
          xml:
            name : ContenuLimite 
        contenuCentral: 
          type: string
          example: exemple contenu central
          xml:
            name : ContenuCentral 
        contenuExclu: 
          type: string
          example: exemple exclusions
          xml:
            name :  ContenuExclu                                                  
      xml:
        name: Poste
    Postes:
      type: array
      description: Tableau représentant les postes d'une nomenclature
      items:
        $ref: '#/components/schemas/Poste'
      xml:
        wrapped: true
        name: Postes
    Nomenclature:
      type: object
      description: Objet représentant une nomenclature
      properties:
        code:
          type: string
          example: 'nafr2'
          xml:
            attribute: true
        uri:
          type: string
          example: 'http://id.insee.fr/codes/nafr2/naf'
          xml:
            attribute: true
        intitule:
          type: string
          example: Nomenclature d'activités française - NAF rév. 2
          xml:
            name: Intitule      
      xml:
        name: Nomenclature
    Nomenclatures:
      type: array
      description: Tableau représentant la liste des nomenclatures
      items:
        $ref: '#/components/schemas/Nomenclature'
      xml:
        wrapped: true
        name: Nomenclatures
    Correspondance:
      type: object
      description: Objet représentant une correspondance
      properties:
        code:
          type: string
          example: '0811Z1301020'
          xml:
            attribute: true
        uri:
          type: string
          example: 'http://dvrmessnczlht01/datalift/codes/emb2018/produit/0811Z1301020'
          xml:
            attribute: true
        intitule:
          type: string
          example: Blanc de craie (à usage industriel)
          xml:
            attribute: true
        codeCible:
          type: string
          example: '08.11Z'
          xml:
            attribute: true
        uriCible:
          type: string
          example: 'http://dvrmessnczlht01/datalift/codes/nafr2/sousClasse/08.11Z'
          xml:
            attribute: true
        intituleCible:
          type: string
          example: Extraction de pierres ornementales et de construction, de calcaire industriel, de gypse, de craie et d'ardoise2
          xml:
            name: IntituleCible
      xml:
        name: Correspondance
    Correspondances:
      type: array
      description: Tableau représentant la liste des correspondances
      items:
        $ref: '#/components/schemas/Correspondance'
      xml:
        wrapped: true
        name: Correspondances        
...
    DescriptionTableCorrespondance:
      type: object
      description: Objet représentant la définition d'une table de correspondance entre deux nomclatures
      xml:
        name: DescriptionTableCorrespondance
      properties:
        idTableCorrespondance:
          type: string
          example: nafr2-cpfr21
          xml:
            attribute: true
        uriTableCorrespondance:
          type: string
          example: 'http://rdf.insee.fr/codes/nafr2-cpfr21'
          xml:
            attribute: true
        idNomclatureSource:
          type: string
          example: 'nafr2'
          xml:
            attribute: true   
        idNomclatureCible:
          type: string
          example: 'cpfr21'
          xml:
            attribute: true                      
        intituleFr:
          type: string
          example: Table de correspondance Naf rév. 2 / Cpf rév. 2.1
          xml:
            name: IntituleFr
        descriptionFr:
          type: string
          example: ''
          xml:
            name: DescriptionFr
    DescriptionsTablesCorrespondance:
      type: array
      description: Tableau représentant des tables de correspondance entre deux nomclatures
      items:
        $ref: '#/components/schemas/DescriptionTableCorrespondance'
      xml:
        wrapped: true
        name: DescriptionsTablesCorrespondance       